backend/requirements/base.txt
Django>=4.2
djangorestframework
psycopg2-binary
celery
redis
python-dotenv
aiohttp>=3.9
beautifulsoup4>=4.12
lxml
